LIBITD-2598. djangosaml2 setup for CAS/SAML authentication
Added the "djangosaml2" package to support authenticating via CAS.

The application uses the “Handle-Administrator” group in Grouper to
determine which users have access.

One significant change (compared to "umd-lib/grove" on which these
configuration changes are based) is that the "AssertionConsumerService"
URL had to be changed from the "djangosaml2" default to "saml2/acs/" to
"users/auth/saml/callback", because that is the callback URL provided
to DIT in the service provider XML file that enables the umd-handle
application to retrieve SAML attributes (CAS requires an exact match
of the endpoint, otherwise it shows an error page).

Added "src/umd_handle/auth.py" to enables user authorization/permissions
to be determined by membership in the "Handle-Administrator" Grouper
group (following the example in "umd-lib/grove").

https://umd-dit.atlassian.net/browse/LIBITD-2598