LIBITD-2601. Modified "jwt_list_tokens" to include "created" timestamp
In keeping with the format from the Rails-based "umd-handle"
application, modified the "jwt_list_tokens" Django management task to
include the "created" timestamp for the tokens.

The Rails-based timestamp had a slightly different presentation for
the timestamp, i.e.:

```
2024-04-23 17:44:01 UTC
```

instead of the Django version:

```
2025-11-11 19:01:30.341155+00:00
```

but it is not believed that there are any applications that rely on
a particular format for this field.

https://umd-dit.atlassian.net/browse/LIBITD-2601