LIBITD-2604. Filter "pathlib.PurePath" deprecation warning from pytest
When running "pytest" the following deprecation warning is printed, and
the test results are marked in yellow (instead of green):

```
DeprecationWarning: pathlib.PurePath.as_uri() is deprecated and scheduled for removal in Python 3.19
```

The deprecated source code is in the "xmlschema" (v2.5.1) package, which
is a transitive dependency of the "pysaml2" package. Since there is no
straightforward way for us to fix this, filtering out the warning to
keep the noise to a minimum.

https://umd-dit.atlassian.net/browse/LIBITD-2604