LIBITD-2616. Replace Django's "URLValidator" with a custom URL validator
Replaced the use of Django’s “URLValidator” with a custom “validate_url”
method that uses Python’s “urllib.parse” method to parse the URL.
The validator then verifies that the URL has either an “http” or “https”
scheme, and that it was a “netloc” (a hostname).

This is necessaru because when requesting a new handle, Avalon may
send a “url” parameter that contains a colon immediately after the
hostname, i.e.,

```
https://av.sandbox.lib.umd.edu:/media_objects/vq27zn67m
```

The Rails-based legacy “umd-handle” application accepted these as
valid URLs.

https://umd-dit.atlassian.net/browse/LIBITD-2616